failureinjection: use processMap for other failures
The processMap type was added in the VM reset failure
to easily capture and stop/restart any running processes.
This is useful for failures that expect to kill all processes
on a given node or cluster and want to easily restart them
without explicitly specifying them all.

This change extracts the processMap type so it can be used by
all failure modes, specifically disk stalls which need to
stop and restart entire clusters.
